feat(vote): 네이티브 시트용 단독 라우트(/vote) + vote_sheet 3단 분기
방향 A — 웹은 투표 UI 디자인만 소유하고, 딤·슬라이드업·바깥탭·닫기는
네이티브 바텀시트가 담당한다. 시트가 웹뷰로 /vote?sheet=1 을 연다.

- VoteContent 신설: 목록/상세/투표/재투표 4상태와 상태 로직을 통째로 옮겼다.
  bare 면 카드 chrome(테두리·라운드·그림자·최대높이) 없이 콘텐츠만,
  onClose 를 안 주면 헤더 X 와 "확인" 버튼도 그리지 않는다(닫기는 앱 몫)
- VoteFloatingWidget 은 컨테이너 전용으로 축소 — 위치·카드 chrome·FAB
- /vote 라우트 + VoteSheetScreen. 투표를 마쳐도 이동하지 않고 결과가 남는다.
  auth 는 tri-state 로 읽어 콜드 스타트에서 로그인 안내가 번쩍이지 않게 한다
- 3단 분기: vote_sheet → 웹 FAB·패널 모두 숨김(앱이 시트로 처리) /
  vote_fab 만 → 기존 FAB 위임 / 둘 다 없음 → 웹 FAB(하위호환)
- vote_sheet 쉘에서도 vote_fab announce 는 계속 보낸다 — 앱이 뱃지를 그리려면
  웹의 visible/count 가 필요하다. 웹은 렌더만 하지 않는다
- native:vote_sheet_closed 수신 시 today 쿼리 무효화 → 배경 화면의 미참여
  수·뱃지 동기화
- /vote 를 우측 레일 제외 목록에 추가
