+++ src/app.component/layout/AppLayoutShell.tsx
@@ -42,6 +42,9 @@ const RIGHT_RAIL_HIDDEN_ROUTES = [
   '/terms',
   '/privacy',
   '/install',
+  // 네이티브 시트가 웹뷰로 여는 투표 단독 라우트 — 시트 안에 레일이 낄 자리가
+  // 없고, 브라우저로 직접 열어도 맨 패널 옆 레일은 어색하다.
+  '/vote',
 ];
 
 // 바텀 네비는 최상위 4개 탭에서만 노출한다. 그 외 서브/상세 화면은
